requirements.txt
requests>=2.31.0,<3
python-dotenv>=1.0.1,<2
flask>=2.3.0,<4
flask-cors>=4.0.0,<5
gunicorn>=21.0.0,<22

# Development and testing dependencies
pytest>=7.0.0,<8
pytest-mock>=3.0.0,<4
